+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 2.1.0
+
+### Added
+- Added Bootstrap/Bootswatch v4.5.2
+
+### Fixed
+- Fixed setting default layout, when none specified.
+
 ## 2.0.0
 
 ### Changed